tflite_test2/requirements.txt
absl-py==2.1.0
apturl==0.5.2
astunparse==1.6.3
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
fasteners==0.14.1
filelock==3.6.0
flatbuffers==24.3.25
future==0.18.2
gast==0.5.4
google-pasta==0.2.0
grpcio==1.63.0
h5py==3.11.0
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
keras==3.3.3
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.3.2
monotonic==1.6
more-itertools==8.10.0
namex==0.0.8
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
olefile==0.46
opencv-python==4.9.0.80
opt-einsum==3.3.0
optree==0.11.0
packaging==21.3
paramiko==2.9.3
pbr==5.8.0
pexpect==4.8.0
Pillow==9.0.1
platformdirs==2.5.1
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
rich==13.7.1
screen-resolution-extra==0.0.0
SecretStorage==3.3.1
six==1.16.0
ssh-import-id==5.11
stevedore==3.5.0
systemd-python==234
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.37.0
termcolor==2.4.0
tflite-runtime==2.14.0
typing_extensions==4.11.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
virtualenv==20.13.0+ds
virtualenv-clone==0.3.0
virtualenvwrapper==4.8.4
wadllib==1.3.6
Werkzeug==3.0.3
wrapt==1.16.0
xdg==5
xkit==0.0.0
zipp==1.0.0
